The weather is a package: component, config and producer side by side
A pure move, following the building package of #578. hisim/components/weather.py
and hisim/components/weather_calculation.py become hisim/components/weather/ with
weather.py (the Weather component), config.py (WeatherConfig, LocationEnum and the
weather_identity sizing fact) and calculation.py (the producer, moved byte for byte;
ARTIFACT_KIND stays "weather_series"). The __init__ re-exports every public name --
Weather, WeatherConfig, LocationEnum, WeatherDataSourceEnum, WeatherSeriesInputs,
WeatherSourceFiles, ARTIFACT_KIND, get_coordinates, produce_weather_series and
calculate_direct_normal_irradiance_in_watt_per_square_meter -- so no importer of
hisim.components.weather had to change.

WeatherDataSourceEnum stays in calculation.py and config.py imports it from there,
rather than the other way round. It is key material, so the producer needs it, and
the producer's import closure is hashed into every weather cache key: hisim.config,
which any configuration module must import, has a closure of 44 modules that reaches
hisim.component, hisim.postprocessing and hisim.sim_repository through its deliberate
lazy imports, all three banned from a producer by spec §12. Measured, not assumed --
ProducerLayering.violations names them. The closure stays at two modules
(hisim.components.weather.calculation and hisim.caching.keys).

Weather.__module__ is pinned to the package in the __init__. get_full_classname and
the energy-system recorder derive a component's class path from __module__, and 33
recorded twins plus the generated JSON schema spell the weather
hisim.components.weather.Weather; without the pin the split would rewrite all of them
to ...weather.weather.Weather, a change of public wire format in a commit whose point
is that nothing changes. It is the one non-mechanical line, and it is why the twins
and the schema are byte-identical below. Its only cost is that
inspect.getsourcelines(Weather) no longer finds the class body, which the code
overview generator already handles; imports, pickling, describe and the JSON class
resolution are unaffected. The one other new line is the runtime-local import of
Weather inside WeatherConfig.get_main_classname, the same backward edge
BuildingConfig resolves the same way.

Checks: pytest tests/test_weather.py tests/test_weather_producer.py
tests/test_weather_identity_sizing.py tests/test_generic_pv_system.py
tests/test_caching_keys.py tests/test_cache_key_portability.py -- 62 passed; the
energy-system, caching, round-trip and singleton suites -- 119 passed; pytest -m base
at its known baseline. record_all_setups.py --check --only basic_household
household_heatpump_building_sizer: every committed twin is exactly what recording
produces. hisim energy-system schema leaves the schema file unchanged;
hisim energy-system describe hisim.components.weather.Weather resolves. Golden pair
basic_household / one_week_60s: GOLDEN CHECK OK, on a cache hit. pylint
--rcfile pylintrc-critical-only on every moved and changed file: 10.00/10. prospector
on hisim/components/weather/ and on the changed tests: 0 messages. flake8
--max-line-length=120 hisim tests: 602 findings against 603 on main, all of them the
same moved lines under their new names plus the long local-import line that
building/config.py carries too. mypy -p hisim and mypy tests/: clean.
check_config_attrs.py: 116 config classes, no findings.
